+++ language-service/language_service/service/tag/__init__.py
@@ -1,6 +1,6 @@
-from language_service.service.nlp.chinese import tag_chinese
-from language_service.service.nlp.english import tag_english
-from language_service.service.nlp.spanish import tag_spanish
+from language_service.service.tag.chinese import tag_chinese
+from language_service.service.tag.english import tag_english
+from language_service.service.tag.spanish import tag_spanish
 
 
 def tag(language, text):
